Initial changes to accomodate 4.x.x Vertx and edge-common

+++ src/main/java/org/folio/edge/ltiCourses/utils/LtiCoursesOkapiClientFactory.java
@@ -6,7 +6,7 @@ import io.vertx.core.Vertx;
 
 public class LtiCoursesOkapiClientFactory extends OkapiClientFactory {
 
-  public LtiCoursesOkapiClientFactory(Vertx vertx, String okapiURL, long reqTimeoutMs) {
+  public LtiCoursesOkapiClientFactory(Vertx vertx, String okapiURL, int reqTimeoutMs) {
     super(vertx, okapiURL, reqTimeoutMs);
   }
